+++ src/components/@shared/layout/SiteHeader/SiteHeader.tsx
@@ -1,175 +1,33 @@
 'use client';
 
-import { Bell, Check, Menu, X } from 'lucide-react';
 import Link from 'next/link';
 import { usePathname, useRouter } from 'next/navigation';
-import {
-  type Dispatch,
-  type SetStateAction,
-  useEffect,
-  useMemo,
-  useState,
-} from 'react';
+import { useEffect, useMemo, useState } from 'react';
 
 import { getNotices } from '@/api/getNotices';
 import ThemeToggle from '@/components/@shared/ui/ThemeToggle';
-import { NAV_ITEMS, SITE_NAME } from '@/constants/site';
+import { SITE_NAME } from '@/constants/site';
 import { useHasMounted } from '@/hooks/useHasMounted';
 import { getReadNoticeIds, hasNewNotice } from '@/utils/noticePolicy';
 
 import {
-  bellDotClass,
-  bellSpanClass,
-  comingSoonBadgeClass,
   desktopBellLinkClass,
-  desktopComingSoonClass,
-  desktopNavClass,
-  desktopNavLinkClass,
   headerClass,
   logoLinkClass,
-  mobileComingSoonClass,
-  mobileBellLinkClass,
-  mobileMenuButtonClass,
-  mobileNavLinkClass,
-  mobileOverlayClass,
 } from './SiteHeader.styles';
-
-const NoticeBell = ({
-  showNoticeDot,
-  className,
-  prefetch,
-}: {
-  showNoticeDot: boolean;
-  className: string;
-  prefetch: (href: string) => void;
-}) => (
-  <Link
-    href="/notice"
-    title="공지사항"
-    aria-label="공지사항"
-    onMouseEnter={() => prefetch('/notice')}
-    className={className}
-  >
-    <span className={bellSpanClass(showNoticeDot)}>
-      <Bell size={17} strokeWidth={2.2} />
-    </span>
-    {showNoticeDot && <span aria-hidden className={bellDotClass} />}
-  </Link>
-);
-
-const DesktopNav = ({
-  pathname,
-  prefetch,
-}: {
-  pathname: string;
-  prefetch: (href: string) => void;
-}) => (
-  <nav className={desktopNavClass}>
-    {NAV_ITEMS.map((item) => {
-      if (item.comingSoon) {
-        return (
-          <span key={item.href} className={desktopComingSoonClass}>
-            {item.label}
-            <span className={comingSoonBadgeClass}>준비중</span>
-          </span>
-        );
-      }
-      const active = pathname === item.href;
-      return (
-        <Link
-          key={item.href}
-          href={item.href}
-          onMouseEnter={() => prefetch(item.href)}
-          className={desktopNavLinkClass(active)}
-        >
-          {item.label}
-        </Link>
-      );
-    })}
-  </nav>
-);
-
-const MobileControls = ({
-  menuOpen,
-  setMenuOpen,
-  showNoticeDot,
-  prefetch,
-}: {
-  menuOpen: boolean;
-  setMenuOpen: Dispatch<SetStateAction<boolean>>;
-  showNoticeDot: boolean;
-  prefetch: (href: string) => void;
-}) => (
-  <div className="ml-auto flex items-center gap-1 min-[641px]:hidden">
-    <NoticeBell
-      showNoticeDot={showNoticeDot}
-      className={mobileBellLinkClass}
-      prefetch={prefetch}
-    />
-    <ThemeToggle />
-    <button
-      type="button"
-      aria-label={menuOpen ? '메뉴 닫기' : '메뉴 열기'}
-      onClick={() => setMenuOpen((v) => !v)}
-      className={mobileMenuButtonClass}
-    >
-      {menuOpen ? (
-        <X size={20} strokeWidth={2} />
-      ) : (
-        <Menu size={20} strokeWidth={2} />
-      )}
-    </button>
-  </div>
-);
-
-const MobileOverlay = ({
-  menuOpen,
-  pathname,
-  setMenuOpen,
-  prefetch,
-}: {
-  menuOpen: boolean;
-  pathname: string;
-  setMenuOpen: Dispatch<SetStateAction<boolean>>;
-  prefetch: (href: string) => void;
-}) => (
-  <div className={mobileOverlayClass(menuOpen)}>
-    <nav className="mx-auto flex w-[min(880px,calc(100%-40px))] flex-col pt-1 pb-4">
-      {NAV_ITEMS.map((item) => {
-        if (item.comingSoon) {
-          return (
-            <span key={item.href} className={mobileComingSoonClass}>
-              {item.label}
-              <span className={comingSoonBadgeClass}>준비중</span>
-            </span>
-          );
-        }
-        const active = pathname === item.href;
-        return (
-          <Link
-            key={item.href}
-            href={item.href}
-            onClick={() => setMenuOpen(false)}
-            onMouseEnter={() => prefetch(item.href)}
-            className={mobileNavLinkClass(active)}
-          >
-            {item.label}
-            {active && <Check size={16} strokeWidth={2.5} aria-hidden />}
-          </Link>
-        );
-      })}
-    </nav>
-  </div>
-);
+import DesktopNav from './_DesktopNav/DesktopNav';
+import MobileControls from './_MobileControls/MobileControls';
+import MobileOverlay from './_MobileOverlay/MobileOverlay';
+import NoticeBell from './_NoticeBell/NoticeBell';
 
 const SiteHeader = () => {
   const pathname = usePathname();
   const router = useRouter();
   const mounted = useHasMounted();
   const notices = useMemo(() => getNotices(), []);
-  const [readIds, setReadIds] = useState<string[]>([]);
+  const [readIds, setReadIds] = useState<string[] | null>(null);
   const showNoticeDot = useMemo(
-    () => mounted && hasNewNotice(notices, readIds),
+    () => mounted && readIds !== null && hasNewNotice(notices, readIds),
     [mounted, notices, readIds]
   );
   const [scrolled, setScrolled] = useState(false);
